Removed the custom hook and moved it into DAW.tsx
