+++ src/locales/en/common.json
@@ -295,8 +295,8 @@
     "shortcuts.zoomHorizontal": "Zoom (Horizontal)",
     "shortcuts.zoomVertical": "Zoom (Vertical)",
     "shortcuts.escapeEditor": "Escape editor focus",
-    "shortcuts.play":"Play Track",
-    "shortcuts.pause":"Pause Track",
+    "shortcuts.play": "Play Track",
+    "shortcuts.pause": "Pause Track",
     "hardware.ctrl": "Ctrl",
     "hardware.cmd": "Cmd",
     "hardware.shift": "Shift",
